+++ src/main/java/org/thema/graphab/graph/GraphGenerator.java
@@ -760,8 +760,9 @@ public class GraphGenerator {
             info += "\n\n" + bundle.getString("NewGraphDialog.intraPatchCheckBox.text");
         }
         
-        info += "\n\n# nodes : " + getGraph().getNodes().size();
-        info += "\n# edges : " + getGraph().getEdges().size();
+        info += "\n\n# " + bundle.getString("nodes") + " : " + getGraph().getNodes().size();
+        info += "\n# " + bundle.getString("edges") + " : " + getGraph().getEdges().size();
+        info += "\n# " + bundle.getString("components") + " : " + getComponents().size();
         
         return info;
     }